+++ WalletWasabi/Blockchain/Transactions/Summary/ForeignOutput.cs
@@ -1,0 +1,10 @@
+using NBitcoin;
+
+namespace WalletWasabi.Blockchain.Transactions.Summary;
+
+public class ForeignOutput : Output
+{
+	public ForeignOutput(Money amount, BitcoinAddress destinationAddress) : base(amount, destinationAddress)
+	{
+	}
+}
